python-app/requirements.txt

psycopg2-binary
# Flask-SQLAlchemy
# sqlalchemy
pandas
